+++ natural-language-understanding/src/main/java/com/ibm/watson/developer_cloud/natural_language_understanding/v1/model/Usage.java
@@ -19,15 +19,16 @@ import com.ibm.watson.developer_cloud.service.model.GenericModel;
  */
 public class Usage extends GenericModel {
 
-  /** Number of features used in the API call. */
-  private Integer features;
+  private Long features;
 
   /**
    * Gets the features.
    *
+   * Number of features used in the API call.
+   *
    * @return the features
    */
-  public Integer getFeatures() {
+  public Long getFeatures() {
     return features;
   }
 
@@ -36,8 +37,7 @@ public class Usage extends GenericModel {
    *
    * @param features the new features
    */
-  public void setFeatures(final Integer features) {
+  public void setFeatures(final long features) {
     this.features = features;
   }
-
 }
